Remove application code from rbfe module
Move to example script
